Updated Aggregator service.
Removed Maintenance models and services related to them. At the moment I think that everything related to this area can be placed in a separate service. Slightly modified scripts until I find suitable data to fill in. Changed some class names in singular and plural for better understanding of what is written. This service is missing filter handling and basically the filtering system needs to be replaced.