Moved `EventSteram` to the Tools folder.
Now it is called `EventDriven` as well as the architecture. The previous library allowed only to use the Request/Reply pattern which was not necessary in some cases. Now this project allows to implement sending and handling of events and also asynchronous sending of requests and handling of message responses. Also, for now only one broker will be used.